Implementacion permisos en todos los modulos

+++ src/main/java/com/example/demo2/api/v1/local/proyecto1/actions/ActionController.java
@@ -5,15 +5,13 @@ import com.example.demo2.api.v1.local.Utils.ResponseLocal;
 import com.example.demo2.api.v1.local.Utils.UtilsLocal;
 import com.example.demo2.api.v1.local.Utils.UtilsService;
 import com.example.demo2.api.v1.local.proyecto1.actions.adapters.ActionDto;
-
 import com.example.demo2.api.v1.local.proyecto1.logs.LogService;
+import com.example.demo2.api.v1.local.proyecto1.permissions.PermissionService;
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
@@ -28,12 +26,20 @@ public class ActionController {
     @Autowired
     LogService logService;
     
+    @Autowired
+    PermissionService permissionService;
+    
     private String myClassName = ActionController.class.getName();
     
-    @PreAuthorize("hasRole('ROLE_ADMIN') OR hasRole('ROLE_USUARIO') OR hasRole('ROLE_LECTURA')")
+    private final String module = "Actions";
+    
+
     @GetMapping
     public ResponseEntity<?> getAll( HttpServletRequest req ){
         ResponseLocal response = new ResponseLocal( logService );
+        Object permission = permissionService.validate( req, this.module, "getAll", response );
+        if( permission != null ) return new ResponseEntity<>( permission, HttpStatus.FORBIDDEN );
+        
         try {
             Object rows = actionService.getAll();
             HttpStatus httpStatus = response.validateService( rows, 
@@ -58,7 +64,6 @@ public class ActionController {
     }
     
     
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
     @GetMapping( path = "/{id}")
     public ResponseEntity<?> getById(
         @PathVariable("id") Integer id,
@@ -66,6 +71,9 @@ public class ActionController {
     )
     {
         ResponseLocal response = new ResponseLocal( logService );
+        Object permission = permissionService.validate( req, this.module, "getOne", response );
+        if( permission != null ) return new ResponseEntity<>( permission, HttpStatus.FORBIDDEN );
+        
         try {
             Object row = this.actionService.getById(id);
             HttpStatus httpStatus = response.validateService( row, 
@@ -89,8 +97,7 @@ public class ActionController {
         }
     }
 
-    
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
+
     @GetMapping("/query")
     public ResponseEntity<?> getByName(
         @RequestParam("name") String name,
@@ -98,6 +105,9 @@ public class ActionController {
     )
     {
         ResponseLocal response = new ResponseLocal( logService );
+        Object permission = permissionService.validate( req, this.module, "getOne", response );
+        if( permission != null ) return new ResponseEntity<>( permission, HttpStatus.FORBIDDEN );
+        
         try {
             Object row = this.actionService.getByName(name);
             HttpStatus httpStatus = response.validateService( row, 
@@ -122,7 +132,6 @@ public class ActionController {
     }
     
     
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
     @PostMapping()
     public ResponseEntity<?> save(
         @Valid @RequestBody ActionDto actionDto,
@@ -130,6 +139,9 @@ public class ActionController {
         HttpServletRequest req
     ){
         ResponseLocal response = new ResponseLocal( logService );
+        Object permission = permissionService.validate( req, this.module, "save", response );
+        if( permission != null ) return new ResponseEntity<>( permission, HttpStatus.FORBIDDEN );
+        
         if( bindingResult.hasErrors() ){
             response.setError( HttpStatus.BAD_REQUEST.value(), "", "", 
                 bindingResult.getAllErrors(),
@@ -163,7 +175,6 @@ public class ActionController {
     }
 
     
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
     @DeleteMapping( path = "/{id}")
     public ResponseEntity<?> delete(
         @PathVariable("id") Integer id,
@@ -171,10 +182,13 @@ public class ActionController {
     )
     {
         ResponseLocal response = new ResponseLocal( logService );
+        Object permission = permissionService.validate( req, this.module, "delete", response );
+        if( permission != null ) return new ResponseEntity<>( permission, HttpStatus.FORBIDDEN );
+        
         try {
             String message = "";
             Object resDel = this.actionService.delete(id);
-            if( ! UtilsService.isErrorService(resDel) ){ 
+            if( ! UtilsService.isErrorService(resDel) ){
                 message = (String) resDel;
                 resDel = null;
             }
@@ -186,17 +200,6 @@ public class ActionController {
                 req
             );
             return new ResponseEntity<>( response, httpStatus );
-        }
-        catch ( DataAccessException e) {
-            response.setError( HttpStatus.BAD_REQUEST.value(), 
-                "No se pudo eliminar la acción.", 
-                e.getMessage(), 
-                UtilsLocal.emptyErrorList(),
-                this.myClassName, 
-                null, 
-                req
-            );
-            return new ResponseEntity<>( response, HttpStatus.BAD_REQUEST );
         }
         catch (Exception e) {
             response.setError( HttpStatus.BAD_REQUEST.value(),
