+++ src/main/java/com/example/demo2/api/v1/local/proyecto1/permissions/PermissionController.java
@@ -5,15 +5,12 @@ import com.example.demo2.api.v1.local.Utils.ResponseLocal;
 import com.example.demo2.api.v1.local.Utils.UtilsLocal;
 import com.example.demo2.api.v1.local.Utils.UtilsService;
 import com.example.demo2.api.v1.local.proyecto1.logs.LogService;
-import com.example.demo2.api.v1.local.proyecto1.modules.adapters.ModuleDto;
 import com.example.demo2.api.v1.local.proyecto1.permissions.adapters.PermissionDto;
 import javax.servlet.http.HttpServletRequest;
 import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
 
@@ -30,12 +27,16 @@ public class PermissionController {
     
     private String myClassName = PermissionController.class.getName();
     
+    private final String module = "Permissions";
     
-    @PreAuthorize("hasRole('ROLE_ADMIN') OR hasRole('ROLE_USUARIO') OR hasRole('ROLE_LECTURA')")
+   
     @GetMapping
     public ResponseEntity<?> getAll( HttpServletRequest req )
     {
         ResponseLocal response = new ResponseLocal( logService );
+        Object permission = permissionService.validate( req, this.module, "getAll", response );
+        if( permission != null ) return new ResponseEntity<>( permission, HttpStatus.FORBIDDEN );
+        
         try {
             Object rows = permissionService.getAll();
             HttpStatus httpStatus = response.validateService( rows, 
@@ -60,7 +61,6 @@ public class PermissionController {
     }
 
     
-    @PreAuthorize("hasRole('ROLE_ADMIN') OR hasRole('ROLE_USUARIO') OR hasRole('ROLE_LECTURA')")
     @GetMapping( path = "/{id}")
     public ResponseEntity<?> getById(
         @PathVariable("id") Integer id,
@@ -68,6 +68,9 @@ public class PermissionController {
     )
     {
         ResponseLocal response = new ResponseLocal( logService );
+        Object permission = permissionService.validate( req, this.module, "getOne", response );
+        if( permission != null ) return new ResponseEntity<>( permission, HttpStatus.FORBIDDEN );
+        
         try {
             Object row = this.permissionService.getById(id);
             HttpStatus httpStatus = response.validateService( row, 
@@ -91,40 +94,7 @@ public class PermissionController {
         }
     }
     
-/*
-    @PreAuthorize("hasRole('ROLE_USUARIO')")
-    @GetMapping("/query")
-    public ResponseEntity<?> getByAction(
-        @RequestParam("name") Integer action_id,
-        HttpServletRequest req
-    )
-    {
-        ResponseLocal response = new ResponseLocal( logService );
-        try {
-            Object row = this.permissionService.getByAction_id(action_id);
-            HttpStatus httpStatus = response.validateService( row, 
-                "Permiso obtenido",
-                this.myClassName, 
-                null, 
-                req
-            );
-            return new ResponseEntity<Object>( response, HttpStatus.OK );
-        }
-        catch (Exception e) {
-            response.setError( HttpStatus.BAD_REQUEST.value(), 
-                "No se pudo obtener el permiso", 
-                e.getMessage(), 
-                UtilsLocal.emptyErrorList(),
-                this.myClassName, 
-                null, 
-                req
-            );
-            return new ResponseEntity<>( response, HttpStatus.BAD_REQUEST );
-        }
-    } */
     
-    
-    @PreAuthorize("hasRole('ROLE_ADMIN') OR hasRole('ROLE_USUARIO')")
     @PostMapping()
     public ResponseEntity<?> save( 
         @Valid @RequestBody PermissionDto permissionDto,
@@ -133,6 +103,8 @@ public class PermissionController {
     )
     {
         ResponseLocal response = new ResponseLocal( logService );
+        Object permission = permissionService.validate( req, this.module, "save", response );
+        if( permission != null ) return new ResponseEntity<>( permission, HttpStatus.FORBIDDEN );
         
         if( bindingResult.hasErrors() ){
             response.setError( HttpStatus.BAD_REQUEST.value(), "", "", 
@@ -167,7 +139,6 @@ public class PermissionController {
     }
 
     
-    @PreAuthorize("hasRole('ROLE_ADMIN')")
     @DeleteMapping( path = "/{id}")
     public ResponseEntity<?> delete(
         @PathVariable("id") Integer id,
@@ -175,6 +146,9 @@ public class PermissionController {
     )
     {
         ResponseLocal response = new ResponseLocal( logService );
+        Object permission = permissionService.validate( req, this.module, "delete", response );
+        if( permission != null ) return new ResponseEntity<>( permission, HttpStatus.FORBIDDEN );
+        
         try {
             String message = "";
             Object resDel = this.permissionService.delete(id);
@@ -190,17 +164,6 @@ public class PermissionController {
                 req
             );
             return new ResponseEntity<>( response, httpStatus );
-        }
-        catch ( DataAccessException e) {
-            response.setError( HttpStatus.BAD_REQUEST.value(), 
-                "No se pudo eliminar el permiso.", 
-                e.getMessage(), 
-                UtilsLocal.emptyErrorList(),
-                this.myClassName, 
-                null, 
-                req
-            );
-            return new ResponseEntity<>( response, HttpStatus.BAD_REQUEST );
         }
         catch (Exception e) {
             response.setError( HttpStatus.BAD_REQUEST.value(),
